% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbca_simple.R
\name{dbca_simple}
\alias{dbca_simple}
\title{Análisis DBCA con gráficos y exportación}
\usage{
dbca_simple(
  archivo,
  nombre_salida = "Resultados/Resultados.xlsx",
  etiqueta_y = "Valor"
)
}
\arguments{
\item{archivo}{nombre del archivo a analizar}

\item{nombre_salida}{nombre del archivo que guardara los CV, los valores p y letras de significancia}

\item{etiqueta_y}{nombre que tomara el eje y en todos los graficos generados}
}
\value{
Resumen de los tratamientos
}
\description{
Esta función realiza un análisis de varianza (DBCA), calcula medias,
agrupaciones de tratamientos (CLD), coeficiente de variación y genera
gráficos de barras y líneas. Además, exporta los resultados a Excel.
}
\examples{
\dontrun{
dbca_simple("Datos/TETRUR.xlsx", nombre_salida = "Resultados/Resultados.xlsx",
etiqueta_y = "Incidencia \%")
}
}
