% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eficacia_abbot.R
\name{eficacia_abbot}
\alias{eficacia_abbot}
\title{Calcular eficacias con ABBOT}
\usage{
eficacia_abbot(archivo, promediar = FALSE)
}
\arguments{
\item{archivo}{nombre del archivo a analizar}

\item{promediar}{condicional que sirve si se necesita promediar mas alla de las repeticiones}
}
\value{
Eficacias de los tratamientos
}
\description{
Esta función realiza el calculo de las eficacias utilizando la formula HENDERSON-TILTON
}
\examples{
\dontrun{
eficacia_abbot("Datos/TETRUR.xlsx")
}


}
