% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hender_tilton.R
\name{hender_tilton}
\alias{hender_tilton}
\title{Calcular eficacias con HENDERSON-TILTON}
\usage{
hender_tilton(archivo, promediar = FALSE)
}
\arguments{
\item{archivo}{nombre del archivo a analizar}

\item{promediar}{condicional que sirve si se necesita promediar mas alla de las repeticiones}
}
\value{
Eficacias de los tratamientos
}
\description{
Esta función realiza el calculo de las eficacias utilizando la formula HENDERSON-TILTON
}
\examples{
\dontrun{
hender_tilton("Datos/TETRUR.xlsx")
}


}
